requirements.txt
# ─── Core Framework ─────────────────────────────────────────────────────────
streamlit==1.35.0

# ─── HTTP Client (for Ollama API) ──────────────────────────────────────────
requests==2.31.0

# ─── Document Parsing ──────────────────────────────────────────────────────
pdfplumber==0.11.0
PyMuPDF==1.27.2.2
python-docx==1.1.2

# ─── PDF Generation ───────────────────────────────────────────────────────
xhtml2pdf==0.2.17

# ─── Data Handling ─────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ─── Data Visualization ───────────────────────────────────────────────────
plotly==5.22.0
